Latest Update

+++ backtesting.rb
@@ -3,9 +3,23 @@ require_relative 'configuration'
 require_relative 'yahoo_client'
 require 'yaml'
 require 'active_support/all'
+require 'optparse'
 
-POSITIONS_FILE = './test.yml'
-OUTPUT_DIR = './output'
+options = {}
+OptionParser.new do |opts|
+  opts.banner = "Usage: backtesting.rb [options]"
+
+  opts.on('-f', '--filename FILENAME', 'Filename') { |v| options[:filename] = v }
+  opts.on('-o', '--output_dir OUTPUT_DIR', 'Output directory') { |v| options[:output_dir] = v }
+end.parse!
+
+unless options[:filename]
+  $stderr.puts "Error: you must specify the --filename option."
+  exit 1
+end
+
+POSITIONS_FILE = options[:filename]
+OUTPUT_DIR = options[:output_dir] || './output'
 
 FileUtils.rm_rf(OUTPUT_DIR)
 FileUtils.mkdir(OUTPUT_DIR)
